requirements.txt
snakemake
sqlalchemy
schema
mysql
sqlalchemy_repr
icecream
colorama
mariadb
gffutils
sh
docopt
multiplug
snacli
